+++ apps/frontend/src/app/3_organisms/Header/components/BridgeMenuItem/BridgeMenuItem.tsx
@@ -97,18 +97,20 @@ export const BridgeMenuItem: FC<BridgeMenuItemProps> = ({ dataAttribute }) => {
       dataAttribute={dataAttribute}
     >
       <Menu className="rounded-t-none rounded-b px-2 py-3 lg:rounded lg:p-1">
-        <MenuItem
-          key={t('header.nav.bridges.subMenu.ethBridge')}
-          text={t('header.nav.bridges.subMenu.ethBridge')}
-          label={
-            !isMobile && t('header.nav.bridges.subMenu.ethBridgeDescription')
-          }
-          dataAttribute={`dapp-menu-ethBridge`}
-          className={classNames('no-underline', {
-            hidden: isRskChain(chainId),
-          })}
-          onClick={handleEthClicked}
-        />
+        {isRskChain(chainId) && (
+          <MenuItem
+            key={t('header.nav.bridges.subMenu.ethBridge')}
+            text={t('header.nav.bridges.subMenu.ethBridge')}
+            label={
+              !isMobile && t('header.nav.bridges.subMenu.ethBridgeDescription')
+            }
+            dataAttribute={`dapp-menu-ethBridge`}
+            className={classNames('no-underline', {
+              hidden: isRskChain(chainId),
+            })}
+            onClick={handleEthClicked}
+          />
+        )}
         <MenuItem
           key={t('header.nav.bridges.subMenu.btcBridge')}
           text={t('header.nav.bridges.subMenu.btcBridge')}