+++ apps/frontend/src/app/3_organisms/TransactionStepDialog/utils.ts
@@ -55,6 +55,11 @@ export const sendOrSimulateTx = async (
   args: string[],
   config: TransactionConfig = {},
 ): Promise<TransactionResponse> => {
+  const from = await request.contract.signer.getAddress();
+  const nonce = await request.contract.provider.getTransactionCount(
+    await request.contract.signer.getAddress(),
+  );
+
   const gasLimit = config.gasLimit ? config.gasLimit?.toString() : undefined;
   const gasPrice = config.gasPrice
     ? parseUnits(config.gasPrice?.toString() || '0', 9)
@@ -92,6 +97,8 @@ export const sendOrSimulateTx = async (
     value: request.value ?? '0',
     gasPrice,
     gasLimit,
+    nonce,
+    from,
   });
 };