+++ apps/frontend/src/app/5_pages/MarketMakingPage/components/AmbientMarketMaking/components/AmbientPoolPositions/hooks/useAmbientPositionValue.ts
@@ -2,7 +2,12 @@ import { useMemo } from 'react';
 
 import { useCurrentChain } from '../../../../../../../../hooks/useChainStore';
 import { useDollarValue } from '../../../../../../../../hooks/useDollarValue';
-import { bigNumberic, decimalic } from '../../../../../../../../utils/math';
+import { findAsset } from '../../../../../../../../utils/asset';
+import {
+  bigNumberic,
+  decimalic,
+  toWei,
+} from '../../../../../../../../utils/math';
 import { AmbientPosition } from '../../../AmbientMarketMaking.types';
 import { AmbientLiquidityPool } from '../../../utils/AmbientLiquidityPool';
 import { useAmbientPositionBalance } from './useAmbientPositionBalance';
@@ -16,12 +21,18 @@ export const useAmbientPositionValue = (
 
   const { usdValue: baseValue } = useDollarValue(
     pool.base,
-    bigNumberic(result?.positionLiqBase || '0').toString(),
+    toWei(
+      bigNumberic(result?.positionLiqBase || '0'),
+      18 - findAsset(pool.base, chainId).decimals,
+    ).toString(),
     chainId,
   );
   const { usdValue: quoteValue } = useDollarValue(
     pool.quote,
-    bigNumberic(result?.positionLiqQuote || '0').toString(),
+    toWei(
+      bigNumberic(result?.positionLiqQuote || '0'),
+      18 - findAsset(pool.quote, chainId).decimals,
+    ).toString(),
     chainId,
   );